requirements.txt
transformers==4.41.2
pillow==10.3.0
python-multipart==0.0.9
torch==2.3.1
fastapi==0.111.0
uvicorn==0.30.1
aiofiles==23.2.1